Date JSON Serialization not working still/User REST API working
